finished lesson 1.2
