+++ src/App.js
@@ -1,34 +1,16 @@
 import React from 'react';
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
-const todoList = [
-  {
-    id: 1,
-    title: "do the dishes",
-  },
-  {
-    id: 2,
-    title: "take out the trash",
-  },
-  {
-    id: 3,
-    title: "sweep the floor",
-  },
-];
+
 
 function App() {
   return (
     <div>
     <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function(item){
-          return ( 
-            <li key ={item.id}>
-              {item.title}
-            </li>
-          );
-        })}
-      </ul>
-      </div>
+    <AddTodoForm/>
+    <TodoList/>
+    </div>
   );
 }
